Update: Smoothing of the SKPP HBL depth, by Patrick, in Noumea, February 2010.
It is a smoothing of the HBL fields from SKPP , with a hanning filter  to avoid noisy signal, especially in case of regional to coarse resolution configurations.  It removes noise and instabilities related, especially near the equator where this noise was critical (Menkes). It conduct to a better behavior of the SST also. This smoothing has not been done yet for the BKPP hbl, it is less critical and it will come later....
This update ask some works on extended range to make it work in OpenMP. It have been tested with OpenMP and MPI, it pass well  the etalon check.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/romsagrif/Roms_tools@584 a0f13d73-952c-0410-bb7c-e0dfd7f2369f