Bug fixed  in 2 way nesting  and MPI parallelization:
- Change in subroutine UpdateGridhis, in zoom.F  : call wrt_grid (ncidhis, hisname, lstr) replace by call wrt_his.
In order to process the writing of the parent updated grid, in history file , in MPI mode parallelization.  

Otherwise, in 2-way nesting and MPI, the update of the parent grid in history NetCDF file at initialisation didn't work.
It was a bit dangerous, because the grid in the his.nc file was the original one, created at the preprocessing phase, the  mask and h arrays  were not  updated by the child grid corresponding.   
In the case of reading the grid in this his file, the mask and bathy were not the correct one and were different from the average file ones.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/romsagrif/Roms_tools@606 a0f13d73-952c-0410-bb7c-e0dfd7f2369f