update set_scoord.F  set_depth.F  and  put_global_atts.F
cpp switch :  NEW_S_COORD

new stretching function for the computation of z_r and z_w.  
hc is no longer limited by hmin and can be arbitrarily large. 
When hc, thetas, thetab are properly chosen, this results
in flat sigma levels between the surface and hc and a higher
vertical resolution at the expected depth of the surface boundary 
layer. When h << hc (i.e. in shallow regions) the vertical resolution 
is uniform.

Warning : the meaning of thetas, thetab and hc is different 
compared to the "old" vertical coordinate.

When the new coordinate is used, an additional global attribute
VertCoordType = 'NEW' is added in the output files.

Reference : 

A.F. Shchepetkin and J.C. McWilliams : A Correction Note for "an 
Forecasting in Terrain- Following Coordinates: Formulation and 
Skill Assessment of the Regional Ocean Modeling System" by Haidvogel 
et al., J. Comp. Phys. 227, pp. 3595-3624


git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/romsagrif/Roms_tools@733 a0f13d73-952c-0410-bb7c-e0dfd7f2369f