BIG BUG  in get_smflx.F  : wrong wind stress rotation in case of rotated grid and direct wind stress direct forcing.  Thanks to X. Capet.
Wind stress rotation are done in ROMSTOOLS, it was done wrongly  a new time Roms_AGRIF code if CURVGRID cpp key defined, and this cpp key is activated by default in REGIONAL configuration !!
We change the CURVGRID by a local cpp key called SMFLUX_CURVGRID in get_smflx.F to avoid this. 
We checked that it is OK  when using BULK_SMFLUX : there is not this wrong extra rotation.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/romsagrif/Roms_tools@1577 a0f13d73-952c-0410-bb7c-e0dfd7f2369f