Major change in the barotropic correction (step3d_uv2.F)
The barotropic correction is stored in nh_ubar and nh_vbar, then the
divergence is subtracted from wz

in the Dirichlet case, the barotropic correction is the normal
one (difference between \int u(z)dz and avg1)

in the Neumann case, the barotropic correction has an additional
contribution: \int du(z)dz, the NH correction

This is a major fix. It fixes the blowup of Grav_adj. The blow-up was
occuring faster as dt was decreased! which is a clear sign of a
coupling problem (the operation is done more often as dt is decreased).