SCRIPTS/Plurimonths_scripts
Thanks to Serena Illig.
Correct initialization of DT.
When using AGRIF, DT is recalculated for the different child grids, but not reset to the value of the parent grid.
The DT time steps are progressively decreased when recalculating the DT for the child grids.


(cherry picked from commit a2aa02168fdcba950a8d2413e4f148c9d10ca8bd)