CROCO_GPU
The big update.
This branch is not supposed to be used :).
It implements a full GPU version of CROCO.
However :
- It has been tested only on the BASIN and REGIONAL/BENGUELA configurations.
- it works only with one single GPU
This branch will need to be updated from the croco-ocean/devs/LA_OMP/croco repository in order to get multi-GPU, multi-nodes capabilities as well as a working GPU non hydrosatic (NBQ) option.