IO : bug fix when using NC4PAR and  DO_NOT_OVERWRITE
When using cppkey DO_NOT_OVERWRITE + NC4PAR , during a restart we get the error message WRT_HIS ERROR while writing variable 'time_step' into history file.
which is related to the NetCDF error: Attempt to extend dataset during NC_INDEPENDENT I/O operation

This is due to the status of the netcdf when creating or opening it with NF_CREATE_PAR or NF_OPEN_PAR (which are used with NC4PAR) being set to nf_independent.
When we create a new file the access is modified in nf_collective but this modification was not done for NF_OPEN_PAR
Basically when the code had to continue to write in a file (without the overwrite) it crashed because it could not because of the bad status (nf_independent instead of nf_collective)

Thanks to Mathieu Le Corre