Some fixes + debug enhancement
step3d_fast.F : Fixes for the NBQ option in the dev2022_GPU_merge brench

initial_nbq.F : missing initialization of thetadiv2_nbq, thetadiv3_nbq

debug.F : Add the logical optional argument 'correct' for the check_tab2d/3d subroutines. In case of a mismatch between two values, if 'correct' is set to true, the code won't stop, but the value of the current run will be replaced by the one read in the check file. Obviously, this should only to be used in special cases.